Moving conditions to Azure::
